manager and userDetail separation of concerns
