+++ src/main/java/bestcommerce/brand/manager/entity/CustomerUserDetails.java
@@ -1,0 +1,52 @@
+package bestcommerce.brand.manager.entity;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+import java.util.stream.Collectors;
+
+@RequiredArgsConstructor
+public class CustomerUserDetails implements UserDetails {
+
+    private final Manager manager;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return this.manager.getRoles().stream()
+                .map(SimpleGrantedAuthority::new)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public String getPassword() {
+        return this.manager.getManagerPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return this.manager.getManagerEmail();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
